requirements.txt
Flask
flask-cors
openai
google-generativeai
elevenlabs
python-dotenv
requests
Pillow
pytesseract
tensorflow
keras
numpy
